zmiana interfejsu eth0 na lo oraz typu timestamp z double na long

+++ location-driver/src/Location.h
@@ -116,7 +116,7 @@ public:
 	double Pos_Y;
 	double Prop;
 	double Pos_Alfa;
-	double timeStamp;
+	long timeStamp;
 
 };
